Moved Tokenizer to syntaxAnalysis and fixed a bunch of imports
