+++ src/Tests/ScannerTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import sw417f20.ebal.Reader.Reader;
-import sw417f20.ebal.Scanner;
-import sw417f20.ebal.Token;
+import sw417f20.ebal.SyntaxAnalysis.Reader;
+import sw417f20.ebal.SyntaxAnalysis.Scanner;
+import sw417f20.ebal.SyntaxAnalysis.Token;
 
 import java.io.BufferedReader;
 import java.io.IOException;
