+++ src/sw417f20/ebal/SyntaxAnalysis/Scanner.java
@@ -1,6 +1,5 @@
-package sw417f20.ebal;
+package sw417f20.ebal.SyntaxAnalysis;
 
-import sw417f20.ebal.SyntaxAnalysis.Reader;
 import java.io.IOException;
 
 public class Scanner {
